+++ src/watts/__init__.py
@@ -5,6 +5,12 @@ from .plugin import *
 from .plugin_openmc import *
 from .plugin_moose import *
 from .plugin_pyarc import *
+from .plugin_sas import *
 from .template import *
 from .parameters import *
 from .database import *
+
+# This allows a user to write watts.Quantity
+from astropy.units import Quantity
+
+__version__ = '0.2.0'