+++ src/watts/parameters.py
@@ -1,39 +1,32 @@
 # SPDX-FileCopyrightText: 2022 UChicago Argonne, LLC
 # SPDX-License-Identifier: MIT
 
-import textwrap
+from __future__ import annotations
+import copy
 from collections import namedtuple
 from collections.abc import MutableMapping, Mapping, Iterable
 from datetime import datetime
 from getpass import getuser
-from typing import Any, Union
+import textwrap
+from typing import Any, Union, BinaryIO
 from warnings import warn
 
-import h5py
+import astropy.units as u
+import dill
 from prettytable import PrettyTable
 
-
-_SAVE_FUNCS = {
-    'set': list
-}
-
-_LOAD_FUNCS = {
-    'tuple': tuple,
-    'list': list,
-    'set': set,
-    'float': float,
-    'int': int,
-    'bool': bool
-}
+# Enable imperial units
+u.imperial.enable()
 
 ParametersMetadata = namedtuple('ParametersMetadata', ['user', 'time'])
+
 
 class Parameters(MutableMapping):
     """User parameters used to generate inputs that are created by plugins
 
     This class behaves like a normal Python dictionary except that it stores
     metadata on (key, value) pairs and provides the ability to save/load the
-    data to an HDF5 file.
+    data to a pickle file.
 
     Attributes
     ----------
@@ -82,6 +75,9 @@ class Parameters(MutableMapping):
 
     def __len__(self):
         return len(self._dict)
+
+    def __repr__(self):
+        return repr(self._dict)
 
     @property
     def warn_duplicates(self) -> bool:
@@ -187,104 +183,86 @@ class Parameters(MutableMapping):
             headers = headers[:2]
         print(table.get_string(fields=headers, sortby=field_to_name[sort_by]))
 
-    def _save_mapping(self, mapping, h5_obj):
-        # Helper function to add metadata
-        def add_metadata(obj, metadata):
-            obj.attrs['user'] = metadata.user
-            obj.attrs['time'] = metadata.time.isoformat()
+    def _save_mapping(self, file_obj: BinaryIO):
+        file_obj.write(dill.dumps(self))
 
-        for key, value in mapping.items():
-            if isinstance(value, Mapping):
-                # If this item is a dictionary, make recursive call
-                group = h5_obj.create_group(key)
-                if isinstance(mapping, type(self)):
-                    add_metadata(group, self._metadata[key])
-                self._save_mapping(value, group)
-            else:
-                # Convert type if necessary. If the type is not listed, return a
-                # "null" function that just returns the original value
-                func = _SAVE_FUNCS.get(type(value).__name__, lambda x: x)
-                file_value = func(value)
-
-                dset = h5_obj.create_dataset(key, data=file_value)
-                dset.attrs['type'] = type(value).__name__
-                if isinstance(mapping, type(self)):
-                    add_metadata(dset, self._metadata[key])
-
-    def save(self, filename_or_obj: Union[str, h5py.Group]):
-        """Save parameters to an HDF5 file/group
+    def save(self, filename_or_obj: Union[str, BinaryIO]):
+        """Save parameters to a pickle file
 
         Parameters
         ----------
         filename_or_obj
-            Path to HDF5 file or HDF5 group object to write to
+            Path to open file or file object write to
         """
         if isinstance(filename_or_obj, str):
-            with h5py.File(filename_or_obj, 'w') as h5file:
-                self._save_mapping(self, h5file)
+            with open(filename_or_obj, 'wb') as file_obj:
+                self._save_mapping(file_obj)
         else:
-            # If HDF5 file/group was passed, use it directly
-            self._save_mapping(self, filename_or_obj)
+            # If file object was passed, use it directly
+            self._save_mapping(filename_or_obj)
 
-    def _load_mapping(self, mapping, h5_obj, root=True):
-        # Helper function to get metadata
-        def metadata_from_obj(obj):
-            user = obj.attrs['user']
-            time = datetime.fromisoformat(obj.attrs['time'])
-            return ParametersMetadata(user, time)
+    def _load_mapping(self, file_obj: BinaryIO):
+        # Load parameters from pickle
+        data = dill.loads(file_obj.read())
 
-        for key, obj in h5_obj.items():
-            if isinstance(obj, h5py.Dataset):
-                # If dataset stores a string type, it needs to be decoded so
-                # that it doesn't return a bytes object
-                if h5py.check_string_dtype(obj.dtype) is not None:
-                    if obj.shape == ():
-                        value = obj[()].decode()
-                    else:
-                        value = obj[()].astype('str')
-                else:
-                    value = obj[()]
+        # Update current instance
+        self.update(data)
+        self._metadata.update(data._metadata)
 
-                # Convert type if indicated. If the type is not listed, return a
-                # "null" function that just returns the original value
-                func = _LOAD_FUNCS.get(obj.attrs['type'], lambda x: x)
-                mapping[key] = func(value)
-
-                if root:
-                    self._metadata[key] = metadata_from_obj(obj)
-
-            elif isinstance(obj, h5py.Group):
-                # For groups, load the mapping recursively
-                mapping[key] = {}
-                if root:
-                    self._metadata[key] = metadata_from_obj(obj)
-
-                self._load_mapping(mapping[key], obj, root=False)
-
-    def load(self, filename_or_obj: Union[str, h5py.Group]):
-        """Load parameters from an HDF5 file
+    def load(self, filename_or_obj: Union[str, BinaryIO]):
+        """Load parameters from a pickle file
 
         Parameters
         ----------
         filename_or_obj
-            Path to HDF5 file or HDF5 group object to read from
+            Path to pickle file or file object to read from
         """
         if isinstance(filename_or_obj, str):
-            with h5py.File(filename_or_obj, 'r') as fh:
-                self._load_mapping(self, fh)
+            with open(filename_or_obj, 'rb') as fh:
+                self._load_mapping(fh)
         else:
-            # If HDF5 file/group was passed, use it directly
-            self._load_mapping(self, filename_or_obj)
+            # If file object was passed, use it directly
+            self._load_mapping(filename_or_obj)
 
     @classmethod
-    def from_hdf5(cls, filename_or_obj: Union[str, h5py.Group]):
-        """Return parameters from HDF5 file/group
+    def from_pickle(cls, filename_or_obj: Union[str, BinaryIO]) -> Parameters:
+        """Return parameters from a pickle file
 
         Parameters
         ----------
         filename_or_obj
-            Path to HDF5 file or HDF5 group object to read from
+            Path to pickle file or file object to read from
         """
         params = cls()
         params.load(filename_or_obj)
+        return params
+
+    def convert_units(self, system: str = 'si', temperature: str = 'K',
+                      inplace: bool = False) -> Parameters:
+        """Perform unit conversion
+
+        Parameters
+        ----------
+        system
+            Desired unit system: 'si' or 'cgs'
+        temperature
+            Desired unit for temperature conversions
+        inplace
+            Whether to modify the parameters (True) or return a copy (False)
+
+        Returns
+        -------
+        A :class:`Parameters` instance with converted units
+        """
+        params = self if inplace else copy.deepcopy(self)
+
+        for key, value in params.items():
+            if isinstance(value, u.Quantity):
+                # Unit conversion for temperature needs to be done separately because
+                # astropy uses a different method to convert temperature.
+                if value.unit.physical_type == 'temperature':
+                    params[key] = value.to(temperature, equivalencies=u.temperature()).value
+                else:
+                    params[key] = getattr(value, system).value
+
         return params