+++ src/watts/template.py
@@ -10,7 +10,16 @@ from .fileutils import PathLike
 from .parameters import Parameters
 
 
-class TemplateModelBuilder:
+class TemplateRenderer:
+    """Helper class for rendering a Jinja template
+
+    Parameters
+    ----------
+    template_file
+        Path to template file
+    **template_kwargs
+        Keywork arguments passed to :class:`jinja2.Template`
+    """
     def __init__(self, template_file: PathLike, **template_kwargs):
         self.template_file = Path(template_file)
         self.template = jinja2.Template(
@@ -20,6 +29,17 @@ class TemplateModelBuilder:
         )
 
     def __call__(self, params: Parameters, filename: Optional[PathLike] = None):
+        """Render the template
+
+        Parameters
+        ----------
+        params
+            User parameters used to fill placeholders
+        filename
+            Filename for rendered template (If none is provided, by default the
+            filename of the template is changed so that the new extension is
+            .rendered)
+        """
         # Default rendered template filename
         if filename is None:
             name = self.template_file.name